pyproject.toml
[project]
name = "chying-agent"
version = "1.5.0"
description = "CHYing Agent - 腾讯云黑客松智能挑战赛"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    # 可观测性
    "langfuse>=3.8.1",
    "langsmith[claude-agent-sdk,otel]>=0.3.0",
    # 核心依赖
    "python-dotenv>=1.2.1",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
    # Claude SDK
    "claude-agent-sdk>=0.1.0",
    # Web API
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.45",
    "python-magic>=0.4.27",
    "httpx-sse>=0.4.3",
    # RAG 知识库
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "rank-bm25>=0.2.2",
    "jieba>=0.42.1",
    "httpx>=0.27.0",
    "mcp>=1.26.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[dependency-groups]
dev = []
